requirements/base.txt
Django~=2.1.6
django-countries~=5.3.3
django-filter~=2.1.0
django-health-check==3.6.1
drf-yasg~=1.11.0
git+https://github.com/Humanitec/django-oauth-toolkit-jwt@v0.5.2#egg=django-oauth-toolkit-jwt
djangorestframework~=3.9.2
psycopg2-binary==2.7.5
